refactor: バックエンドコードを大幅にリファクタリングして構造を改善
## 主要な変更点

### 📁 ファイル構造の最適化
- 不要な開発用ファイルを削除 (db_inspector.py, log_monitor.py)
- モジュール別にディレクトリを整理
  - ai/ : AI関連機能 (Gemini API, メッセージハンドリング)
  - websocket/ : WebSocket関連機能 (接続管理, ハンドリング)
  - utils/ : 共通ユーティリティ (セッション管理)

### 🔧 コードの分離と整理
- websocket.py (348行) → 複数の小さなモジュールに分割
- gemini_api.py → ai/gemini_client.py に移動
- WebSocket関連の7つの関数を適切なモジュールに分離
- セッション管理ロジックを共通ユーティリティとして独立

### 📈 メンテナンス性の向上
- 単一責任原則に準拠した設計
- 明確なモジュール境界の確立
- 型定義の専用ファイル分離
- import文の簡素化

### 🧹 コード削除
- 614行の開発用コードを削除 (db_inspector + log_monitor)
- 重複コードの排除
- 使用されていない関数の削除

### 📊 結果
- ファイル数: 9個 → 13個 (モジュール化)
- 総行数: 約30%削減 (不要コード除去)
- 複雑性: 大幅改善 (関数分離)
- テスタビリティ: 向上 (モジュール分離)
