+++ src/backend/ai/__init__.py
@@ -1,0 +1,1 @@
+"""AI関連モジュール"""
