+++ src/backend/ai/gemini_client.py
@@ -34,7 +34,7 @@ class GeminiAPIClient:
 
     def _load_system_prompt(self) -> None:
         """システムプロンプトを読み込む."""
-        prompt_path = Path(__file__).parent.parent.parent / "prompts" / "001_ハルト.md"
+        prompt_path = Path(__file__).parent.parent.parent.parent / "prompts" / "001_ハルト.md"
         logger.info(f"プロンプトファイル読み込み試行: {prompt_path}")
         try:
             with open(prompt_path, encoding="utf-8") as f:
