+++ src/backend/utils/__init__.py
@@ -1,0 +1,1 @@
+"""共通ユーティリティモジュール"""
