+++ src/backend/websocket/__init__.py
@@ -1,0 +1,6 @@
+"""WebSocket関連モジュール"""
+
+from .handler import handle_websocket_message
+from .manager import ConnectionManager, manager
+
+__all__ = ["handle_websocket_message", "ConnectionManager", "manager"]
