+++ src/backend/websocket/types.py
@@ -1,0 +1,22 @@
+"""WebSocket関連の型定義"""
+
+from typing import TypedDict
+
+
+class WebSocketMessageData(TypedDict):
+    """WebSocketメッセージのdata部分の型定義"""
+
+    id: str
+    channel_id: str
+    user_id: str
+    user_name: str
+    content: str
+    timestamp: str
+    is_own_message: bool
+
+
+class WebSocketMessage(TypedDict):
+    """WebSocketメッセージの型定義"""
+
+    type: str
+    data: WebSocketMessageData | None
